[frontend] modularize input panel
